move compareSemVer to legend-shared
