+++ .changeset/curly-frogs-rush.md
@@ -1,0 +1,6 @@
+---
+"@finos/legend-application-studio": patch
+"@finos/legend-application-query": patch
+"@finos/legend-storage": patch
+"@finos/legend-shared": patch
+---
