+++ .changeset/curly-frogs-rush2.md
@@ -1,0 +1,6 @@
+---
+"@finos/legend-storage": patch
+"@finos/legend-shared": patch
+---
+
+Move semver util to legend-shared
